+++ currencyconverterapi/client.py
@@ -1,29 +1,48 @@
 from json import load
 from urllib.request import urlopen
 
+from .models import PairCache, Settings
+
 
 class CurrencyConverterApi:
 
     API_BASE = "https://free.currconv.com"
 
-    api_key: str
+    __settings: Settings
 
     def __init__(self, api_key: str = "") -> None:
-        assert api_key, "Get free API key from: https://free.currencyconverterapi.com/free-api-key"
-        self.api_key = api_key
+        self.__settings = Settings.from_file()
+        if api_key:
+            self.__settings.cache_api_key(api_key)
+        assert self.__settings.api_key, "Get free API key from: https://free.currencyconverterapi.com/free-api-key"
 
-        self.__currencies: set[str] = set()
+        self.fetch_currencies()
+        self.__settings.save()
 
     def get(self, url):
         with urlopen(url) as f:
             return load(f)
 
     @property
+    def api_key(self) -> str:
+        return self.__settings.api_key
+
+    @property
     def currencies(self) -> set[str]:
-        if not self.__currencies:
+        self.fetch_currencies()
+        return self.__settings.currencies
+
+    @property
+    def cached_pairs(self) -> dict[str, PairCache]:
+        return self.__settings.cached_pairs
+
+    def fetch_currencies(self):
+        if not self.__settings.currencies:
             url = f"{CurrencyConverterApi.API_BASE}/api/v7/currencies?apiKey={self.api_key}"
-            self.__currencies = set(self.get(url)["results"].keys())
-        return self.__currencies
+            self.__settings.currencies = set(self.get(url)["results"].keys())
+
+    def is_cached(self, pair: str) -> bool:
+        return pair in self.cached_pairs and not self.cached_pairs[pair].is_outdated()
 
     def is_exist(self, symbol: str) -> bool:
         return symbol.upper() in self.currencies
@@ -34,5 +53,8 @@ class CurrencyConverterApi:
                 raise ValueError(f"{symbol} is not available currency")
 
         pair = f"{source.upper()}_{destination.upper()}"
-        url = f"{CurrencyConverterApi.API_BASE}/api/v7/convert?q={pair}&compact=ultra&apiKey={self.api_key}"
-        return self.get(url)[pair]
+        if not self.is_cached(pair):
+            print("Fetching...")
+            url = f"{CurrencyConverterApi.API_BASE}/api/v7/convert?q={pair}&compact=ultra&apiKey={self.api_key}"
+            self.__settings.cache_pair(pair, self.get(url)[pair])
+        return self.cached_pairs[pair].value