+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 📋 Changelog
 
+# 🏷 1.0.1
+
+- Added cache support
+- All data is stored in `$HOME/currencyconverterapi.yml`
+- All cached pair value are refreshed every 10 mins
+
 # 🏷 1.0.0
 
 - First initialization